storeInfo bug fix ,  now can change CusNum type between total and detail

+++ app/src/main/java/com/citrus/mCitrusTablet/model/vo/StoreInfo.kt
@@ -26,10 +26,10 @@ data class Info (
     val businessEndHours: String,
 
     @SerializedName("Longitude")
-    val longitude: Long,
+    val longitude: Double,
 
     @SerializedName("Latitude")
-    val latitude: Long,
+    val latitude: Double,
 
     @SerializedName("Note")
     val note: String,
